requirements.txt
asgiref==3.8.1
asttokens==2.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
distlib==0.3.9
Django==5.1.7
et-xmlfile==1.1.0
executing==2.1.0
filelock==3.16.1
fonttools==4.54.1
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mysql-connector-python==9.1.0
mysqlclient==2.2.7
nest-asyncio==1.6.0
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
pyzmq==26.2.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sqlparse==0.5.3
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.2
virtualenv==20.27.1
wcwidth==0.2.13
